+++ tests/test_like_utils.py
@@ -1,4 +1,4 @@
-from flaskbb.core.settings import flaskbb_config
+from flaskbb.settings import flaskbb_config
 from flaskbb.extensions import db
 from flaskbb.forum.models import Forum, Post, Topic
 from sqlalchemy import event